Project/Image_Generator-main-Deployment/requirements.txt
python-dotenv
Flask
gunicorn
requests
googletrans==4.0.0-rc1
